This fixes #111. (Add support for Valve VTF format.)

+++ src/librpbase/TextFuncs_wchar.hpp
@@ -83,7 +83,7 @@ static inline std::string W2U8(const wchar_t *wcs, int len = -1)
 static inline std::string W2U8(const std::wstring &wcs)
 {
 	return LibRpBase::utf16_to_rp_string(
-		reinterpret_cast<const char16_t*>(wcs.data()), wcs.size());
+		reinterpret_cast<const char16_t*>(wcs.data()), (int)wcs.size());
 }
 
 #endif /* __ROMPROPERTIES_LIBRPBASE_TEXTFUNCS_WCHAR_HPP__ */
